refactor: cut over to wos.project, delete wos.validators
Update all callers (scripts/lint.py, tests) to import directly from
wos.project. Add validate_project/check_project_files as module-level
functions alongside Project class — functional API, not a shim.
